requirements.txt
fastapi
uvicorn[standard]
pillow
google-generativeai
google-cloud-storage
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib